close issue #https://github.com/openebs/openebs/issues/206
1. Why is this change necessary ?

This e2e test case is one of the foundational steps
required to automatically calculate time taken
for a jiva controller reschedule. There are some
amount of manual work involved which should be taken
care of by future issues & implementations.

This e2e test case is deployed as a Kubernetes Job
in a Kubernetes setup.

2. How does this change address the issue ?

The main.go:
- creates openebs jiva volume
- kills the pod belonging to jiva volume controller
  - this triggers a reschedule of volume controller
deployment
- logs the time taken for jiva volume controller
reschedule

3. How to verify this change ?

- check the Pod logs that is associated with this job

4. What side effects does this change have ?

- None